created common steps for purpose
